fix: 키워드 post api 상위로 이동
