+++ src/pages/write/Draw.tsx
@@ -2,19 +2,9 @@ import { useOutletContext } from "react-router-dom";
 import dummy from "../../assets/dummy/_Image.png";
 import { FADEINANIMATION } from "../../styles/animations";
 import { ContextProps } from "./Layout/DiaryWriteFlowLayout";
-import { useEffect } from "react";
-import { postKeyword } from "@/api/api";
 
 const Draw = () => {
-  const { diaryId, keywords } = useOutletContext<ContextProps>();
-
-  useEffect(() => {
-    const setKeywordToDiary = async () => {
-      await postKeyword({ diaryId: diaryId, keywords: keywords });
-    };
-
-    if (diaryId !== "0") setKeywordToDiary();
-  }, [diaryId]);
+  const { diaryId } = useOutletContext<ContextProps>();
 
   return (
     <div className="items-centergap-[2rem] flex h-full flex-col justify-center font-Binggrae text-gray-900 dark:text-gray-50">
